fix(Datagrid): case issue fix rename
