+++ packages/ibm-products/src/components/Datagrid/Datagrid/DatagridRow.tsx
@@ -11,7 +11,7 @@ import { px } from '@carbon/layout';
 import { selectionColumnId } from '../common-column-ids';
 import cx from 'classnames';
 import { pkg, carbon } from '../../../settings';
-import { DatagridAILabel } from './addons/AiLabel1/DatagridAiLabel1';
+import { DatagridAILabel } from './addons/AiLabel/DatagridAiLabel';
 import { DataGridState } from '../types';
 
 const blockClass = `${pkg.prefix}--datagrid`;
